CV/requirements.txt
CacheControl==0.14.3
cachetools==5.5.2
certifi==2022.12.7
cffi==2.0.0
charset-normalizer==2.1.1
contourpy==1.3.2
cryptography==45.0.7
cvzone==1.6.1
cycler==0.12.1
filelock==3.13.1
firebase-admin==6.2.0
fonttools==4.59.2
google-api-core==2.25.1
google-api-python-client==2.181.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.3.1
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.74.0
httplib2==0.30.0
idna==3.4
Jinja2==3.1.4
kiwisolver==1.4.9
MarkupSafe==2.1.5
matplotlib==3.7.2
mpmath==1.3.0
msgpack==1.1.1
networkx==3.3
numpy==1.24.3
opencv-python==4.8.1.78
packaging==25.0
pandas==2.3.2
Pillow==10.0.0
proto-plus==1.26.1
protobuf==6.32.0
psutil==7.0.0
psycopg2-binary==2.9.10
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.10.1
pyparsing==3.0.9
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.1
requests==2.31.0
rsa==4.9.1
scipy==1.11.1
seaborn==0.13.2
six==1.17.0
sympy==1.13.3
thop==0.1.1.post2209072238
torch==2.0.1+cpu
torchvision==0.15.2+cpu
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.2
ultralytics==8.0.196
uritemplate==4.2.0
urllib3==1.26.13
